restructured redux to use feature names
