+++ client/src/controllers/EntryCreate/EntryCreate.js
@@ -3,7 +3,7 @@ import './EntryCreate.scss'
 import {useParams} from "react-router-dom";
 
 import {connect} from 'react-redux'
-import {getSingleStoryAction} from '../../redux/actions/db_get'
+import {getSingleStoryAction} from '../../redux/actions/story'
 
 import EntryCreateForm from '../../components/EntryCreateForm/EntryCreateForm'
 import GoHomeButton from '../../components/ButtonTypes/GoHomeButton/GoHomeButton'
